Tinggal benerin struktur data ama PARSING FUNGSI

+++ modusFile/function/function.h
@@ -7,8 +7,8 @@
 	r2 = row 2;
 	c2 = column 2;
  */
-
-int sum (int r1, int c1, int r2, int c2);
+#include "../modusFile.h"
+double sum (int r1, int c1, int r2, int c2);
 
 //fungsi yang lain sama parameternya. r1,c1,r2,c2. tambahin sendiri ya.
